Work on #53
Good news and bad news.

The good news: I have an implementation of enemies damaging the player
that will work for us with some tweaks in other spots.

The bad news: In order for my implementation of damage to work, the
enemy's collider must be a trigger. When that happens, it falls through
the floor unless it is a kinematic object. When that happens, it floats
in mid air, and attaches to the player. When the player jumps, the enemy
appears to jump as well, but ends up jumping infinitely out of the
scene. This feels like an issue that is not a part of anything that I've
been working on, but feel free to correct me if I'm wrong.

Hopefully it's a quick fix, and then I can work on #54. Also, I have a
good idea of how I can implement a debuff/buff system logically, but
I'll cross that bridge when I come to it.

Thanks,
Brad Konsela