Minor change and thoughts
Good Afternoon,

So I know this is a rather inconsequential change, but looking at the
Killable script made me think that using abstraction for the Buff class
would be the best for us in the long run. From there, I'm wondering if
we just attach the script to each object that's going to have a
buff/debuff effect and have damage controller check the object's buff
information and handle damage accordingly in the coroutines. Hope that
makes sense, and we can certainly hash it out in person.

Thanks
Brad Konsela